requirements.txt
Flask==2.3.3
redis==5.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
prometheus-client==0.17.1
etcd3==0.12.0

# Django Stack (推荐)
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-filter==23.4
drf-spectacular==0.26.5
